Derive one kind of rule condition for both halves of cleanup
They were derived differently. A merge suggestion built its condition from the
cluster's reduced stem and the final payee name; the rule tab built it from what
the payee's imports actually share, and only that half knew a date is not a
merchant, that a word used by hundreds of payees names none of them, or that a
plain substring usually does the job. One budget produced two shaped rules
depending on which tab proposed them.

The derivation moves to its own module and both halves use it.

Patterns also read as MAYA.*BORDERS rather than
\\bMAYA[^A-Za-z0-9]*BORDERS\\b. It is broader, which is the point — it also
catches a word appearing between two parts of a name — and a rule nobody can
read at a glance is a rule nobody can check. Anything it over-reaches, the
backtest catches. A plain contains is tried before either.
