+++ docs-site/src/content/docs/user-guide/payee-cleanup.mdx
@@ -149,6 +149,9 @@ they have in common isn't the merchant:
 - **A run that catches every import doesn't automatically win.** One import written `LVLUP` shouldn't
   reduce `LVL UP FITNESS` to `FITNESS`.
 
+The same reasoning produces the rule offered on a merge suggestion, so one budget
+gets one kind of rule wherever it was proposed from.
+
 The pattern is shown on the row, and you can rewrite it: choose the field, choose **matches** or
 **contains**, and type your own. The backtest re-runs as you go, and tells you if it now catches
 nothing, will not compile, or reaches another payee. Your edits survive a re-scan.
