+++ src/features/payee-cleanup/lib/ruleGaps.ts
@@ -32,117 +32,23 @@ import type { PayeeCleanupCandidate } from "../types";
 import { isCleanupEligible } from "./eligibility";
 import { reduceFully } from "./reduce";
 import {
-  buildCandidates,
+  chooseCondition,
   classifyRelatedRules,
   exactNameCoverage,
-  normalizePatternText,
-  rankCandidates,
   scoreCandidate,
   type CandidateScore,
   type ImportedTextRow,
   type RuleCandidate,
   type SourceField,
 } from "./ruleCandidates";
-
-/**
- * How far the ladder may shorten a core. `LVL UP` and `FITNESS` are too little
- * to hang a rule on; `EMIRATES` and `READY SET` are enough. It does not apply to
- * the run itself — see `coreLadder`.
- */
-const MIN_CORE_LENGTH = 8;
-
-/**
- * The longest a merchant core may be.
- *
- * Bank transfer records are mostly rigid boilerplate — an IBAN, a reference, a
- * branch code — and all of it is constant across transfers to the same payee, so
- * the longest shared run is nearly the whole record. Four words is enough to
- * name any merchant, and the backtest still has to agree the result is safe.
- */
-const MAX_CORE_TOKENS = 4;
-
-/**
- * How much better one run's coverage must be to outweigh a worse-looking core.
- * Below this the two are treated as equally well supported, and the one that
- * reads more like the merchant's name wins.
- */
-const COVERAGE_MARGIN = 0.15;
-
-/**
- * Hashtags are markers, never merchant text: `#2026-05` is a date and `#API` is
- * a channel. Left in, they corrupt the core — two payees whose imports happen to
- * fall in the same month share `05`, which is how `Green Planet` became
- * `05 GREEN PLANET`. Stripped before the core is derived, never from the text the
- * rule is matched against.
- */
-function withoutMarkers(text: string): string {
-  return text.replace(/#\S+/g, " ");
-}
-
-function coreTokens(text: string): string[] {
-  return normalizePatternText(withoutMarkers(text)).split(" ").filter(Boolean);
-}
-
-/**
- * How many payees each word turns up for.
- *
- * The budget's own answer to "is this word a merchant, or is it scenery?" —
- * `MUDON` belongs to one payee, `DUBAI` to dozens. No list of cities is needed,
- * and it works for any country: the same reasoning the cleanup scan uses to
- * learn a bank's boilerplate, applied a word at a time.
- */
-export type TokenSpread = { payeesFor: Map<string, number>; payeeCount: number };
-
-export function measureTokenSpread(rows: ImportedTextRow[]): TokenSpread {
-  const seen = new Map<string, Set<string>>();
-  for (const row of rows) {
-    if (!row.payeeId) continue;
-    for (const token of new Set(coreTokens(row.text))) {
-      const payees = seen.get(token) ?? new Set<string>();
-      payees.add(row.payeeId);
-      seen.set(token, payees);
-    }
-  }
-
-  const payeesFor = new Map<string, number>();
-  for (const [token, payees] of seen) payeesFor.set(token, payees.size);
-  return {
-    payeesFor,
-    payeeCount: new Set(rows.map((r) => r.payeeId).filter(Boolean)).size,
-  };
-}
-
-/**
- * How much a run looks like this payee's name rather than scenery.
- *
- * A word counts for it, a word shared with many other payees counts against, and
- * anything carrying a digit counts against. Without the first two, the longest
- * shared run wins on length alone: Grammarly got a rule keyed on the phone
- * number in its statement line, and a builder got `LA ROSA` — the street its
- * invoices mention — instead of its own name.
- */
-function runQuality(tokens: string[], spread: TokenSpread | undefined): number {
-  const generic = Math.max(2, Math.ceil((spread?.payeeCount ?? 0) * 0.05));
-
-  let score = 0;
-  for (const token of tokens) {
-    if (/\d/.test(token)) {
-      score -= 1;
-    } else if (/^[A-Z]{2,}$/.test(token)) {
-      score += (spread?.payeesFor.get(token) ?? 1) > generic ? -1 : 1;
-    }
-  }
-  return score;
-}
-
-/** The rarest word in a run, for choosing between two equally good ones. */
-function rarestToken(tokens: string[], spread: TokenSpread | undefined): number {
-  let rarest = Number.MAX_SAFE_INTEGER;
-  for (const token of tokens) {
-    rarest = Math.min(rarest, spread?.payeesFor.get(token) ?? 1);
-  }
-  return rarest;
-}
+import {
+  commonTokenRun,
+  coreTokens,
+  followedInSomeText,
+  maximalCommonRun,
+  measureTokenSpread,
+  type TokenSpread,
+} from "./core";
 
 /** The share of a payee's transactions an existing rule must catch to settle it. */
 const COVERED_SHARE = 0.5;
@@ -657,7 +563,14 @@ function proposeRule(
     const longest = maximalCommonRun(uncoveredTexts, weights);
     const boundaryShown = longest !== null && followedInSomeText(longest, uncoveredTexts);
 
-    const best = chooseCondition(run, field, inputs.rows, payee, !boundaryShown);
+    const best = chooseCondition(
+      run,
+      field,
+      inputs.rows,
+      new Set([payee.id]),
+      new Set([payee.name.toUpperCase()]),
+      !boundaryShown
+    );
     if (!best || best.expectedMatches === 0) return null;
 
     return { shape: "matches", field, candidate: best.candidate, score: best, extendsRule: null };
@@ -676,234 +589,6 @@ function proposeRule(
     texts: recurring.map((row) => row.text),
     extendsRule,
   };
-}
-
-/**
- * The longest run of words that *most* of the import text contains.
- *
- * This is what makes a rule catch text it has never seen. Requiring the texts to
- * *reduce* to the same stem was too strict — the reducer is tuned for payee
- * names and leaves different leading noise on each one — but so was requiring
- * the run in every single text. One outlier is enough to ruin it: nine imports
- * reading `LVL UP FITNESS CTR DUBAI UAE` and one reading `LVLUP FITNESS` share
- * only `FITNESS`, and a stray `EMIRATES62385176881` means the payee's other
- * fifteen `EMIRATES` imports share nothing at all.
- *
- * So the run has to cover a **majority of the transactions**, not every distinct
- * string — weighted by transaction count, because a one-off oddity should not
- * outvote text that arrives every month.
- *
- * Contiguous by design: `buildCandidates` joins the words with "any run of
- * non-alphanumerics", which only means anything if they were adjacent.
- */
-export function commonTokenRun(
-  texts: string[],
-  weights?: number[],
-  minShare = 0.5,
-  spread?: TokenSpread
-): string | null {
-  const entries = texts
-    .map((text, i) => ({ tokens: coreTokens(text), weight: weights?.[i] ?? 1 }))
-    .filter((entry) => entry.tokens.length > 0);
-  if (entries.length === 0) return null;
-
-  const total = entries.reduce((sum, entry) => sum + entry.weight, 0);
-  if (total === 0) return null;
-
-  const containsRun = (tokens: string[], run: string[]) => {
-    for (let i = 0; i + run.length <= tokens.length; i++) {
-      if (run.every((word, k) => tokens[i + k] === word)) return true;
-    }
-    return false;
-  };
-
-  // Seeded from the text carrying the most transactions, since that is the one
-  // most likely to be representative.
-  const seed = [...entries].sort((a, b) => b.weight - a.weight)[0].tokens;
-
-  let best: {
-    run: string[];
-    coverage: number;
-    quality: number;
-    rarest: number;
-  } | null = null;
-
-  for (let length = Math.min(seed.length, MAX_CORE_TOKENS); length >= 1; length--) {
-    for (let start = 0; start + length <= seed.length; start++) {
-      const run = seed.slice(start, start + length);
-      const matching = entries.filter((entry) => containsRun(entry.tokens, run));
-      // Shared by at least two texts when there is more than one, or a single
-      // dominant string defines the run all by itself.
-      if (entries.length > 1 && matching.length < 2) continue;
-
-      const coverage =
-        matching.reduce((sum, entry) => sum + entry.weight, 0) / total;
-      if (coverage < minShare) continue;
-
-      const quality = runQuality(run, spread);
-      const rarest = rarestToken(run, spread);
-
-      // Coverage first, but only when it is *materially* better. A run catching
-      // every import beats one catching half — `LIFE 29 PHY` over
-      // `LIFE 29 PHY 1264 DUBAI`. A run catching every import does not beat one
-      // catching nine in ten if that costs the merchant's name: one payee
-      // writing `LVLUP` once must not reduce `LVL UP FITNESS` to `FITNESS`.
-      const candidate = { run, coverage, quality, rarest };
-
-      // Anything that reads like a name beats anything that reads like scenery,
-      // whatever the coverage. Generic words have high coverage *because* they
-      // are generic — `DUBAI` closes more of this payee's imports than
-      // `EMIRATES` does, and a rule built on it would catch half the budget.
-      const namelike = quality > 0;
-      const bestNamelike = best !== null && best.quality > 0;
-
-      const better =
-        best === null ||
-        (namelike !== bestNamelike
-          ? namelike
-          : // Then coverage, but only when materially better: a run catching
-            // every import must not beat one catching nine in ten if that costs
-            // the merchant's name.
-            coverage > best.coverage + COVERAGE_MARGIN ||
-            (coverage > best.coverage - COVERAGE_MARGIN &&
-              (quality > best.quality ||
-                (quality === best.quality &&
-                  (rarest < best.rarest ||
-                    (rarest === best.rarest && run.length > best.run.length))))));
-
-      if (better) best = candidate;
-    }
-  }
-
-  if (!best) return null;
-  const joined = best.run.join(" ");
-  // A laxer floor than the trimming one: `COLES` and `IKEA` are whole merchants,
-  // and the evidence says so. `MIN_CORE_LENGTH` stops the ladder shortening
-  // *past* the evidence; it does not overrule it.
-  return best.run.length >= 2 || best.run[0].length >= 4 ? joined : null;
-}
-
-/** Whether any text carries more words after this run. */
-function followedInSomeText(run: string[], texts: string[]): boolean {
-  return texts.some((text) => {
-    const tokens = coreTokens(text);
-    for (let i = 0; i + run.length <= tokens.length; i++) {
-      if (run.every((word, k) => tokens[i + k] === word)) {
-        return i + run.length < tokens.length;
-      }
-    }
-    return false;
-  });
-}
-
-/**
- * The longest run these texts share, with no cap and no judgement about how it
- * reads. Used only to ask where the shared text stops.
- */
-function maximalCommonRun(texts: string[], weights: number[]): string[] | null {
-  const entries = texts
-    .map((text, i) => ({ tokens: coreTokens(text), weight: weights[i] ?? 1 }))
-    .filter((entry) => entry.tokens.length > 0);
-  if (entries.length === 0) return null;
-
-  const total = entries.reduce((sum, entry) => sum + entry.weight, 0);
-  const seed = [...entries].sort((a, b) => b.weight - a.weight)[0].tokens;
-
-  for (let length = seed.length; length >= 1; length--) {
-    for (let start = 0; start + length <= seed.length; start++) {
-      const run = seed.slice(start, start + length);
-      const matching = entries.filter((entry) => {
-        for (let i = 0; i + run.length <= entry.tokens.length; i++) {
-          if (run.every((word, k) => entry.tokens[i + k] === word)) return true;
-        }
-        return false;
-      });
-      // Shared, not merely present in the heaviest string. Without this a
-      // dominant text is its own longest run, so nothing follows it and every
-      // core looked safe to shorten.
-      if (entries.length > 1 && matching.length < 2) continue;
-
-      const covered = matching.reduce((sum, entry) => sum + entry.weight, 0);
-      if (covered / total >= 0.5) return run;
-    }
-  }
-  return null;
-}
-
-/**
- * Shorter and shorter leading parts of the run, longest first.
- *
- * The run is the longest core the history *permits*, not the shortest that
- * *works*. Three identical `Google Storage Mountain View CA SAR10.99` imports
- * make the price look like part of the merchant, and a rule carrying it breaks
- * the day the price changes.
- */
-function coreLadder(run: string): string[] {
-  const tokens = run.split(" ");
-  // The full run is always allowed however short it is — it is what the imports
-  // actually share. Only the shortened forms have to clear the floor.
-  const ladder: string[] = [run];
-  for (let length = tokens.length - 1; length >= 1; length--) {
-    const core = tokens.slice(0, length).join(" ");
-    if (core.length >= MIN_CORE_LENGTH) ladder.push(core);
-  }
-  return ladder;
-}
-
-/**
- * The simplest condition that catches this payee and nothing else.
- *
- * Two preferences, in order:
- *
- * 1. **The shortest safe core.** Shortening can only ever match *more*, so the
- *    shortest core the backtest still clears is the one most likely to catch a
- *    variant that has not arrived yet — which is the entire point of the rule.
- *    It also drops trailing noise, like a subscription price, without needing to
- *    know what it is.
- * 2. **`contains` over a pattern.** `contains "READY SET GO KIDS"` and
- *    `matches \bREADY[^A-Za-z0-9]*SET[^A-Za-z0-9]*GO[^A-Za-z0-9]*KIDS\b` do the
- *    same job, and only one of them can be read at a glance. The pattern is kept
- *    for what it exists for: text whose punctuation varies between imports,
- *    where a literal substring would miss.
- */
-function chooseCondition(
-  run: string,
-  field: SourceField,
-  rows: ImportedTextRow[],
-  payee: PayeeCleanupCandidate,
-  trimmable: boolean
-): CandidateScore | null {
-  const clusterIds = new Set([payee.id]);
-  const clusterNames = new Set([payee.name.toUpperCase()]);
-  // Only shorten when the imports never showed where the merchant ends. If
-  // something follows the run in some text — `READY SET GO KIDS` then `AMUS` in
-  // one import and `DUBAI` in another — the data located the boundary and
-  // second-guessing it would throw away the evidence.
-  const ladder = trimmable ? coreLadder(run) : coreLadder(run).slice(0, 1);
-  if (ladder.length === 0) return null;
-
-  let longestAttempt: CandidateScore | null = null;
-
-  // Shortest first, so the first safe one wins.
-  for (const core of [...ladder].reverse()) {
-    const scored = buildCandidates(core, field).map((candidate) =>
-      scoreCandidate(candidate, rows, clusterIds, clusterNames)
-    );
-    const safe = scored.filter((s) => s.unexpectedMatches === 0 && s.expectedMatches > 0);
-
-    if (safe.length === 0) {
-      // Keep the longest core's best attempt: a payee whose text cannot be
-      // caught safely still deserves a proposal and an explanation rather than
-      // disappearing from the list.
-      if (core === ladder[0]) longestAttempt = rankCandidates(scored)[0] ?? null;
-      continue;
-    }
-
-    const simplest = safe.find((s) => s.candidate.op === "contains");
-    return simplest ?? rankCandidates(safe)[0] ?? null;
-  }
-
-  return longestAttempt;
 }
 
 /**
